src/checker.py
class CommentChecker(object):
    def __init__(self, max_length: int = 1000):
        """
        CommentChecker is a class that has several methods to check if a comment is valid.

        Args:
        - max_length: the maximum length of a comment
        """
        self.max_length = max_length

    def has_alpha(self, text: str):
        """
        Checks if the comment has at least one letter.

        Args:
        - text: the comment to be checked

        Returns:
        - True if the comment has at least one letter, False otherwise
        """
        return any(char.isalpha() for char in text)

    def only_contains_anon(self, text: str):
        """
        Checks if the comment only contains anonymous reserved words.

        Args:
        - text: the comment to be checked

        Returns:
        - True if the comment only contains anonymous reserved words, False otherwise
        """
        anon_words = ["USER", "HASHTAG", "URL"]
        for word in text.split():
            if word not in anon_words:
                return False
        return True

    def has_acceptable_length(self, text: str):
        """
        Checks if the comment has an acceptable length.

        Args:
        - text: the comment to be checked

        Returns:
        - True if the comment has an acceptable length, False otherwise
        """
        return len(text) <= self.max_length

    def is_empty(self, text: str):
        """
        Checks if the comment is empty.
            
        Args:
        - text: the comment to be checked

        Returns:
        - True if the comment is empty, False otherwise
        """
        if text in ["", " ", "\n", "\t", None]:
            return True
        else:
            return False